standard input/output methods
read integer/byte, write integer/byte/string, helper methods, fixed static allocation, static data size

+++ src/main/java/yapl/impl/ByteUtils.java
@@ -23,7 +23,7 @@ public class ByteUtils {
 
     public static byte[] numberAsBytesArray(short i) {
         ByteBuffer bb = ByteBuffer.allocate(2);
-        bb.putInt(i);
+        bb.putShort(i);
         return bb.array();
     }
